+++ torch_geometric/graphgym/logger.py
@@ -1,6 +1,9 @@
 import logging
 import math
 import sys
+import time
+import warnings
+from typing import Any, Dict, Optional
 
 import torch
 
@@ -9,6 +12,16 @@ from torch_geometric.graphgym import register
 from torch_geometric.graphgym.config import cfg
 from torch_geometric.graphgym.utils.device import get_current_gpu_usage
 from torch_geometric.graphgym.utils.io import dict_to_json, dict_to_tb
+
+try:
+    import pytorch_lightning as pl
+    from pytorch_lightning import Callback
+
+except ImportError:
+    pl = None
+    Callback = object
+    warnings.warn("Please install 'pytorch_lightning' for using the GraphGym "
+                  "experiment manager via 'pip install pytorch_lightning'")
 
 
 def set_printing():
@@ -236,14 +249,125 @@ def infer_task():
 
 
 def create_logger():
-    """
-    Create logger for the experiment
-
-    Returns: List of logger objects
-
-    """
+    r"""Create logger for the experiment."""
     loggers = []
     names = ['train', 'val', 'test']
     for i, dataset in enumerate(range(cfg.share.num_splits)):
         loggers.append(Logger(name=names[i], task_type=infer_task()))
     return loggers
+
+
+class LoggerCallback(Callback):
+    def __init__(self):
+        self._logger = create_logger()
+        self._train_epoch_start_time = None
+        self._val_epoch_start_time = None
+        self._test_epoch_start_time = None
+
+    @property
+    def train_logger(self) -> Any:
+        return self._logger[0]
+
+    @property
+    def val_logger(self) -> Any:
+        return self._logger[1]
+
+    @property
+    def test_logger(self) -> Any:
+        return self._logger[2]
+
+    def _get_stats(
+        self,
+        epoch_start_time: int,
+        outputs: Dict[str, Any],
+        trainer: 'pl.Trainer',
+    ) -> Dict:
+        return dict(
+            true=outputs['true'].detach().cpu(),
+            pred=outputs['pred_score'].detach().cpu(),
+            loss=float(outputs['loss']),
+            lr=trainer.lr_scheduler_configs[0].scheduler.get_last_lr()[0],
+            time_used=time.time() - epoch_start_time,
+            params=cfg.params,
+        )
+
+    def on_train_epoch_start(
+        self,
+        trainer: 'pl.Trainer',
+        pl_module: 'pl.LightningModule',
+    ):
+        self._train_epoch_start_time = time.time()
+
+    def on_validation_epoch_start(
+        self,
+        trainer: 'pl.Trainer',
+        pl_module: 'pl.LightningModule',
+    ):
+        self._val_epoch_start_time = time.time()
+
+    def on_test_epoch_start(
+        self,
+        trainer: 'pl.Trainer',
+        pl_module: 'pl.LightningModule',
+    ):
+        self._test_epoch_start_time = time.time()
+
+    def on_train_batch_end(
+        self,
+        trainer: 'pl.Trainer',
+        pl_module: 'pl.LightningModule',
+        outputs: Dict[str, Any],
+        batch: Any,
+        batch_idx: int,
+        unused: int = 0,
+    ) -> None:
+        stats = self._get_stats(self._train_epoch_start_time, outputs, trainer)
+        self.train_logger.update_stats(**stats)
+
+    def on_validation_batch_end(
+        self,
+        trainer: 'pl.Trainer',
+        pl_module: 'pl.LightningModule',
+        outputs: Optional[Dict[str, Any]],
+        batch: Any,
+        batch_idx: int,
+        dataloader_idx: int,
+    ) -> None:
+        stats = self._get_stats(self._val_epoch_start_time, outputs, trainer)
+        self.val_logger.update_stats(**stats)
+
+    def on_test_batch_end(
+        self,
+        trainer: 'pl.Trainer',
+        pl_module: 'pl.LightningModule',
+        outputs: Optional[Dict[str, Any]],
+        batch: Any,
+        batch_idx: int,
+        dataloader_idx: int,
+    ) -> None:
+        stats = self._get_stats(self._test_epoch_start_time, outputs, trainer)
+        self.test_logger.update_stats(**stats)
+
+    def on_train_epoch_end(
+        self,
+        trainer: 'pl.Trainer',
+        pl_module: 'pl.LightningModule',
+    ):
+        self.train_logger.write_epoch(trainer.current_epoch)
+        self.train_logger.close()
+
+    def on_validation_epoch_end(
+        self,
+        trainer: 'pl.Trainer',
+        pl_module: 'pl.LightningModule',
+    ):
+        self.val_logger.write_epoch(trainer.current_epoch)
+        self.val_logger.close()
+
+    def on_test_epoch_end(
+        self,
+        trainer: 'pl.Trainer',
+        pl_module: 'pl.LightningModule',
+    ):
+        self.test_logger.write_epoch(trainer.current_epoch)
+        self.test_logger.close()